Fix: Make filters extend `PySteppableBase`
Transformed filters created by calling either `transform` or `apply` did
not inherit from `PySteppableBase`, even though the underlying filters
did. This caused them to be not registered when calling
`ConfigBuilder.setup()`, since they were not instantiating a Steppable.
Fix this (for now) by making every filter inherit from
`PySteppableBase`. This should be fine for the moment, since filters
are meant as a wrapper around the querying done in the steppables
anyways.