Fix: PBC in com tracker
When tracking the COM of clusters, the com_tracker plugin did not
account for the periodic boundary conditions correctly. Fix this by
unwrapping the positions when tracking the COM.
Add the `wrap_box` parameter to the com_tracker, which allows the user
to specify the simulation box manually. If specified, wrap the recorded
COM around the specified coordinates to restore the previous behaviour.